+++ R/sim.coor.sing-data.R
@@ -8,7 +8,7 @@
 #' @format \describe{
 #'   
 #'   \item{sim.coor.sing}{Simulated coordinated singing events that overlap and do not overlap most of the time,
-#'   for use with \code{coor_test}}
+#'   for use with \code{test_coordination}}
 #'   
 #'   }
 #'