+++ R/spectrograms.R
@@ -81,7 +81,7 @@
 #' version of \code{\link[seewave]{spectro}}, so it takes the same arguments. 
 #' @return Image files containing spectrograms of the signals listed in the input data frame.
 #' @family spectrogram creators
-#' @seealso \code{\link{freq_track}} for creating spectrograms to visualize 
+#' @seealso \code{\link{track_freq_contour}} for creating spectrograms to visualize 
 #'   frequency measurements by \code{\link{spectro_analysis}}, \code{\link{snr_spectrograms}} for 
 #'   creating spectrograms to optimize noise margins used in \code{\link{sig2noise}}
 #' @export